Duc v1

+++ back-end/blood-donation-support-system/src/main/java/com/swp391/bloodcare/dto/BloodDTO.java
@@ -14,7 +14,7 @@ public class BloodDTO {
     private Blood.BloodType bloodType;
     private Blood.RhFactor rhFactor;
     private Boolean isRareBlood;
-    private Integer quantity;
+    private long quantity;
     private String bloodMatch;
 
     public BloodDTO(Blood blood) {
