+++ back-end/blood-donation-support-system/src/main/java/com/swp391/bloodcare/entity/Blood.java
@@ -35,8 +35,7 @@ public class Blood {
     private Boolean isRareBlood;
 
     @Column(name = "quantity")
-    private Integer quantity;
-
+    private long quantity;
 
     @Column(name ="blood_match")
     private String bloodMatch;
